+++ src/Sentry/Internal/Hub.cs
@@ -68,6 +68,8 @@ internal class Hub : IHub, IDisposable
             PushScope();
         }
 
+        Logger = SentryStructuredLogger.Create(this, options, _clock);
+
 #if MEMORY_DUMP_SUPPORTED
         if (options.HeapDumpOptions is not null)
         {
@@ -794,6 +796,8 @@ internal class Hub : IHub, IDisposable
         _memoryMonitor?.Dispose();
 #endif
 
+        Logger.Dispose();
+
         try
         {
             CurrentClient.FlushAsync(_options.ShutdownTimeout).ConfigureAwait(false).GetAwaiter().GetResult();
@@ -818,4 +822,7 @@ internal class Hub : IHub, IDisposable
     }
 
     public SentryId LastEventId => CurrentScope.LastEventId;
+
+    [Experimental(DiagnosticId.ExperimentalFeature)]
+    public SentryStructuredLogger Logger { get; }
 }